[front] fix: keep the pod databases client off the sandbox server module
PodDatabasesTab imported MAX_TABLE_ROWS_PAGE_SIZE from
lib/api/sandbox_functions/dsbx_db, which reaches for node:crypto, Redis and the
sandbox lifecycle. That pulled the whole server module into the browser graph
and made the pod page fail to load with a ChunkLoadError.

The wire types and the page-size cap now live in types/api/sandbox/pod_databases
(client-safe); dsbx_db and db_list import them from there.
